@@ -51,30 +51,28 @@
     "karma-firefox-launcher": "^0.1.4",
     "karma-mocha": "^0.1.10",
     "karma-sourcemap-loader": "^0.3.4",
-    "karma-webpack": "^1.7.0",
-    "less": "^2.4.0",
-    "less-loader": "^2.2.0",
-    "lodash": "^3.10.1",
-    "mocha": "^2.2.1",
-    "modernizr": "^3.0.0-alpha.4",
-    "raw-loader": "^0.5.1",
-    "react": "^0.13.1",
-    "react-bootstrap": "^0.20.0",
-    "react-highlight": "^0.4.1",
-    "react-hot-loader": "^1.2.5",
-    "react-router": "^0.13.2",
-    "source-map": "^0.5.2",
-    "style-loader": "^0.9.0",
+    "karma-webpack": "^1.7",
+    "less": "^2.4",
+    "less-loader": "^2.2",
+    "lodash": "^3.10",
+    "mocha": "^2.2",
+    "modernizr": "^3.0",
+    "raw-loader": "^0.5",
+    "react": "^0.14.0",
+    "react-addons-css-transition-group": "^0.14.0",
+    "react-bootstrap": "^0.27.2",
+    "react-dom": "^0.14.0",
+    "react-highlight": "^0.4",
+    "react-hot-loader": "^1.2",
+    "react-router": "^0.13",
+    "source-map": "^0.5",
+    "style-loader": "^0.13",
     "supports-color": "^3.1.2",
     "text-loader": "0.0.1",
-    "url-loader": "^0.5.5",
-    "webpack": "^1.12.2",
-    "webpack-dev-middleware": "^1.2.0",
-    "webpack-dev-server": "^1.11.0"
-  },
-  "peerDependencies": {
-    "react": "*",
-    "lodash": "*"
+    "url-loader": "^0.5",
+    "webpack": "^1.12",
+    "webpack-dev-middleware": "^1.2",
+    "webpack-dev-server": "^1.11"
   },
   "dependencies": {}
 }
