@@ -18,7 +18,7 @@ var CSS = require("../css");
 var ReactTransitionEvents = require("react/lib/ReactTransitionEvents");
 
 var onlyChild = require("react/lib/onlyChild");
-var warning = require("react/lib/warning");
+var warning = require("fbjs/lib/warning");
 
 // We don't remove the element from the DOM until we receive an animationend or
 // transitionend event. If the user screws up and forgets to add an animation
